+++ CHANGELOG.md
@@ -7,3 +7,12 @@
 * Add nullable mechanism
 * Add global constants e.g `__<TYPE_NAME>__`
 * Rewrite/upgrade tests
+
+
+## 2.1
+* Add `guard` synonyms to `hardCheck` methods:
+```
+TypeChecker::hardCheckInt(123) => TypeChecker::guardInt(123);
+TypeChecker::hardCheckBool(false) => TypeChecker::guardBool(false);
+....
+```